+++ cmake/thirdparty/BLTSetupOpenMP.cmake
@@ -1,4 +1,4 @@
-# Copyright (c) 2017-2023, Lawrence Livermore National Security, LLC and
+# Copyright (c) 2017-2024, Lawrence Livermore National Security, LLC and
 # other BLT Project Developers. See the top-level LICENSE file for details
 # 
 # SPDX-License-Identifier: (BSD-3-Clause)
@@ -13,7 +13,7 @@ find_package(OpenMP REQUIRED)
 # check if the openmp flags used for C/C++ are different from the openmp flags
 # used by the Fortran compiler
 set(BLT_OPENMP_FLAGS_DIFFER FALSE CACHE BOOL "")
-if (ENABLE_FORTRAN)
+if (BLT_ENABLE_FORTRAN)
   string(COMPARE NOTEQUAL "${OpenMP_CXX_FLAGS}" "${OpenMP_Fortran_FLAGS}"
          BLT_OPENMP_FLAGS_DIFFER )
 endif()
@@ -24,12 +24,12 @@ endif()
 set(_compile_flags ${OpenMP_CXX_FLAGS})
 set(_link_flags  ${OpenMP_CXX_FLAGS})
 
-if(NOT COMPILER_FAMILY_IS_MSVC AND ENABLE_CUDA AND BLT_OPENMP_FLAGS_DIFFER)
+if(NOT COMPILER_FAMILY_IS_MSVC AND BLT_ENABLE_CUDA AND BLT_OPENMP_FLAGS_DIFFER)
     set(_compile_flags
         $<$<AND:$<NOT:$<COMPILE_LANGUAGE:CUDA>>,$<NOT:$<COMPILE_LANGUAGE:Fortran>>>:${OpenMP_CXX_FLAGS}> 
         $<$<COMPILE_LANGUAGE:CUDA>:-Xcompiler=${OpenMP_CXX_FLAGS}>
         $<$<COMPILE_LANGUAGE:Fortran>:${OpenMP_Fortran_FLAGS}>)
-elseif(NOT COMPILER_FAMILY_IS_MSVC AND ENABLE_CUDA)
+elseif(NOT COMPILER_FAMILY_IS_MSVC AND BLT_ENABLE_CUDA)
     set(_compile_flags
         $<$<NOT:$<COMPILE_LANGUAGE:CUDA>>:${OpenMP_CXX_FLAGS}> 
         $<$<COMPILE_LANGUAGE:CUDA>:-Xcompiler=${OpenMP_CXX_FLAGS}>)
@@ -56,3 +56,5 @@ blt_import_library(NAME openmp
                    COMPILE_FLAGS ${_compile_flags}
                    LINK_FLAGS    ${_link_flags}
                    EXPORTABLE    ${BLT_EXPORT_THIRDPARTY})
+
+add_library(blt::openmp ALIAS openmp)